improve commands from twitch - now able to execute multiple commands on each event
