+++ server/twitch.handler.ts
@@ -3,22 +3,11 @@ import {EmoteObj, Options} from 'tmi.js';
 import {Subscription} from 'rxjs';
 import {PersistenceInstance} from './persistence';
 import {startWith} from 'rxjs/operators';
-import {
-  Dictionary,
-  Twitch,
-  TwitchConfig,
-  TwitchEventTypes,
-  TwitchTriggerCommand
-} from '@memebox/contracts';
+import {Dictionary, Twitch, TwitchConfig, TwitchEventTypes, TwitchTriggerCommand} from '@memebox/contracts';
 import {triggerMediaClipById} from './websocket-server';
 import {Logger} from 'winston';
 import {newLogger} from './logger.utils';
-
-declare module 'tmi.js' {
-  export interface Badges {
-    founder?: string;
-  }
-}
+import {getCommandsOfMessage, getLevelOfTags} from "./twitch.functions";
 
 export class TwitchHandler {
   private tmiClient: tmi.Client;
@@ -91,66 +80,74 @@ export class TwitchHandler {
     this.tmiClient.on('message', (channel, tags, message, self) => {
       if (self) return;
 
+      // This is the bot handler, nothing after that needs to be handled
       if (this.twitchConfig.bot?.enabled && this.twitchConfig.bot.auth && message === '!commands') {
         this.handleCommandsRequest(tags, message);
 
         return false;
       }
 
-      const command = this.getCommandOfMessage(message, TwitchEventTypes.message);
+      const foundCommandsIterator = getCommandsOfMessage(this.twitchSettings, message, TwitchEventTypes.message);
 
-      this.log({
-        type: 'message',
-        tags,
-        message
-      });
+      for (const command of foundCommandsIterator) {
+        this.log({
+          type: 'message',
+          tags,
+          message
+        });
 
-      this.handle({
-        // event: TwitchEventTypes.message,
-        message,
-        command,
-        tags
-      });
+        this.handle({
+          // event: TwitchEventTypes.message,
+          message,
+          command,
+          tags
+        });
+      }
     });
 
     this.tmiClient.on('cheer', (channel, tags, message) => {
-      const command = this.getCommandOfMessage(message, TwitchEventTypes.bits, {
+      const foundCommandsIterator = getCommandsOfMessage(this.twitchSettings, message, TwitchEventTypes.bits, {
         amount: parseInt(tags.bits)
       });
 
-      this.log({
-        type: 'cheer',
-        tags,
-        message
-      });
+      for (const command of foundCommandsIterator) {
+        this.log({
+          type: 'cheer',
+          tags,
+          message
+        });
 
-      // todo add the correct twitchevent-types!
-      this.handle({
-        // event: TwitchEventTypes.message,
-        message,
-        command,
-        tags
-      });
+        // todo add the correct twitchevent-types!
+        this.handle({
+          // event: TwitchEventTypes.message,
+          message,
+          command,
+          tags
+        });
+      }
     });
 
     this.tmiClient.on('raided', (channel: string, username: string, viewers: number) => {
-      const command = this.getCommandOfMessage('', TwitchEventTypes.raid, {
+      const foundCommandsIterator = getCommandsOfMessage(this.twitchSettings,
+        '', TwitchEventTypes.raid, {
         amount: viewers
       });
 
-      this.log({
-        type: 'raid',
-        username,
-        viewers
-      });
+      for (const command of foundCommandsIterator) {
+        this.log({
+          type: 'raid',
+          username,
+          viewers
+        });
 
-      // todo add the correct twitchevent-types!
-      this.handle({
-        // event: TwitchEventTypes.message,
-        message: '',
-        command,
-        tags: {}
-      });
+        // todo add the correct twitchevent-types!
+        this.handle({
+          // event: TwitchEventTypes.message,
+          message: '',
+          command,
+          tags: {}
+        });
+      }
     });
 
     this.subscribeToUnusedTmiEvents();
@@ -194,7 +191,7 @@ export class TwitchHandler {
   }
 
   handleCommandsRequest(tags: tmi.ChatUserstate, message: string): void {
-    const foundLevels = this.getLevelOfTags(tags);
+    const foundLevels = getLevelOfTags(tags);
     const commands = this.twitchSettings.filter((event) => {
       const trigger: TwitchTriggerCommand = {message, command: event, tags};
 
@@ -218,83 +215,6 @@ export class TwitchHandler {
       .catch(console.error);
   }
 
-  getCommandOfMessage(
-    message: string,
-    event: TwitchEventTypes,
-    eventOptions?: TwitchEventOptions
-  ): Twitch {
-    if (!message && !eventOptions) {
-      return null;
-    }
-
-    let foundCommand: Twitch = null;
-    for (const twitchSetting of this.twitchSettings) {
-      if (!twitchSetting.active) {
-        continue;
-      }
-
-      if (eventOptions && event === twitchSetting.event) {
-        const minAmount = twitchSetting.minAmount || 0;
-        const maxAmount = twitchSetting.maxAmount || Infinity;
-
-        if (eventOptions.amount >= minAmount && eventOptions.amount <= maxAmount) {
-          return twitchSetting;
-        }
-      }
-
-      if (twitchSetting.event !== TwitchEventTypes.message) {
-        continue;
-      }
-
-      if (message.toLowerCase().includes(twitchSetting.contains.toLowerCase())) {
-        if (!foundCommand) {
-          foundCommand = twitchSetting;
-        } else {
-          // another command, example
-          //!party
-          //!partyhard
-          // always take the "longer" command
-
-          if (foundCommand.contains.length < twitchSetting.contains.length) {
-            foundCommand = twitchSetting;
-          }
-        }
-      }
-    }
-
-    return foundCommand;
-  }
-
-  getLevelOfTags(userState: tmi.Userstate): string[] {
-    const levels = ['user'];
-
-    if (!userState.badges) {
-      return levels;
-    }
-
-    if (userState.badges.broadcaster) {
-      levels.push('broadcaster');
-    }
-
-    if (userState.badges.moderator) {
-      levels.push('moderator');
-    }
-
-    if (userState.badges.founder) {
-      levels.push('founder', 'subscriber');
-    }
-
-    if (userState.badges.subscriber) {
-      levels.push('subscriber');
-    }
-
-    if (userState.badges.vip) {
-      levels.push('vip');
-    }
-
-    return levels;
-  }
-
   handle(trigger: TwitchTriggerCommand) {
     if (trigger.command) {
       this.log(`Trigger "${trigger.command.name}" Type - ${trigger.command.event}`);
@@ -307,7 +227,7 @@ export class TwitchHandler {
         tags: trigger.tags
       });
 
-      const foundLevels = this.getLevelOfTags(trigger.tags);
+      const foundLevels = getLevelOfTags(trigger.tags);
       const allowedToTrigger = this.isAllowedToTrigger(trigger, foundLevels);
 
       if (allowedToTrigger) {
@@ -395,8 +315,4 @@ export class TwitchHandler {
       });
     });
   }
-}
-
-interface TwitchEventOptions {
-  amount: number
 }
